+++ tests/browser/wait_for_request.test.js
@@ -24,17 +24,17 @@ describe("Wait For Request", function() {
         setTimeout(() => {
             browser.clickText("click me");
         }, 100);
-        await browser.assert.request.url(/api/).exactly(0);
+        await browser.assert.requests.url(/api/).exactly(0);
         await browser.requests.waitForRequest("http://localhost:3456/api");
-        await browser.assert.request.url(/api/);
+        await browser.assert.requests.url(/api/);
     });
 
     it("Wait For Request Already Made", async() => {
         await browser.clickText("click me");
         await browser.wait(10);
-        await browser.assert.request.url(/api/).exactly(1);
+        await browser.assert.requests.url(/api/).exactly(1);
         await browser.requests.waitForRequest("http://localhost:3456/api");
-        await browser.assert.request.url(/api/).exactly(1);
+        await browser.assert.requests.url(/api/).exactly(1);
     });
 
     it("Wait For Request With Mock", async() => {
@@ -42,9 +42,9 @@ describe("Wait For Request", function() {
         setTimeout(() => {
             browser.clickText("click me");
         }, 100);
-        await browser.assert.request.url(/api/).exactly(0);
+        await browser.assert.requests.url(/api/).exactly(0);
         await browser.requests.waitForRequest("http://localhost:3456/api");
-        await browser.assert.request.url(/api/);
+        await browser.assert.requests.url(/api/);
     });
 
     it("Wait For Response With Mock", async() => {
@@ -54,10 +54,10 @@ describe("Wait For Request", function() {
         });
         await browser.clickText("click me");
         await browser.wait(10);
-        await browser.assert.request.url(/api/).exactly(1);
-        await browser.assert.request.url(/api/).responseBody("test").exactly(0);
+        await browser.assert.requests.url(/api/).exactly(1);
+        await browser.assert.requests.url(/api/).responseBody("test").exactly(0);
         await browser.requests.waitForResponse("http://localhost:3456/api", 1000);
-        await browser.assert.request.url(/api/).responseBody("test");
+        await browser.assert.requests.url(/api/).responseBody("test");
     });
 
     it("Wait For Response Already Made", async() => {
@@ -66,10 +66,10 @@ describe("Wait For Request", function() {
         });
         await browser.clickText("click me");
         await browser.wait(100);
-        await browser.assert.request.url(/api/).exactly(1);
-        await browser.assert.request.url(/api/).responseBody("test").exactly(1);
+        await browser.assert.requests.url(/api/).exactly(1);
+        await browser.assert.requests.url(/api/).responseBody("test").exactly(1);
         await browser.requests.waitForResponse("http://localhost:3456/api");
-        await browser.assert.request.url(/api/).responseBody("test").exactly(1);
+        await browser.assert.requests.url(/api/).responseBody("test").exactly(1);
     });
 
     it("Wait For Request Timeout", async() => {
@@ -88,15 +88,15 @@ describe("Wait For Request", function() {
         setTimeout(() => {
             browser.clickText("click me");
         }, 100);
-        await browser.assert.request.url(/api/).exactly(0);
+        await browser.assert.requests.url(/api/).exactly(0);
         await browser.requests.waitForNextRequest("http://localhost:3456/api");
-        await browser.assert.request.url(/api/);
+        await browser.assert.requests.url(/api/);
     });
 
     it("Wait For Next Request Already Made", async() => {
         await browser.clickText("click me");
         await browser.wait(100);
-        await browser.assert.request.url(/api/).exactly(1);
+        await browser.assert.requests.url(/api/).exactly(1);
         await utils.assertThrowsAsync(async() => {
             await browser.requests.waitForNextRequest("http://localhost:3456/api", 10);
         }, `TimeoutError: [waitForNextRequest] Waiting for request "http://localhost:3456/api", timeout of 10ms exceeded.`);
@@ -109,10 +109,10 @@ describe("Wait For Request", function() {
         });
         await browser.clickText("click me");
         await browser.wait(100);
-        await browser.assert.request.url(/api/).exactly(1);
-        await browser.assert.request.url(/api/).responseBody("test").exactly(0);
+        await browser.assert.requests.url(/api/).exactly(1);
+        await browser.assert.requests.url(/api/).responseBody("test").exactly(0);
         await browser.requests.waitForNextResponse("http://localhost:3456/api", 1000);
-        await browser.assert.request.url(/api/).responseBody("test");
+        await browser.assert.requests.url(/api/).responseBody("test");
     });
 
     it("Wait For Next Response Already Made", async() => {
@@ -121,11 +121,11 @@ describe("Wait For Request", function() {
         });
         await browser.clickText("click me");
         await browser.wait(100);
-        await browser.assert.request.url(/api/).exactly(1);
-        await browser.assert.request.url(/api/).responseBody("test").exactly(1);
+        await browser.assert.requests.url(/api/).exactly(1);
+        await browser.assert.requests.url(/api/).responseBody("test").exactly(1);
         await utils.assertThrowsAsync(async() => {
             await browser.requests.waitForNextResponse("http://localhost:3456/api", 10);
         }, `TimeoutError: [waitForNextResponse] Waiting for response "http://localhost:3456/api", timeout of 10ms exceeded.`);
-        await browser.assert.request.url(/api/).responseBody("test").exactly(1);
+        await browser.assert.requests.url(/api/).responseBody("test").exactly(1);
     });
 });